Refactoring. Impl movabs for x86_64.

+++ src/interpreter/mod.rs
@@ -3,14 +3,14 @@ pub mod exec;
 use crate::parser::ast::Val;
 
 pub trait Execution {
-   fn exec(&self) -> Val;
+    fn exec(&self) -> Val;
 }
 
 #[cfg(test)]
 mod test {
-    use crate::parser::ast::{Val, parse_exp};
-    use crate::parser::lexer::Lexer;
     use crate::interpreter::Execution;
+    use crate::parser::ast::{parse_exp, Val};
+    use crate::parser::lexer::Lexer;
 
     fn perform(input: &str, result: Val) {
         let mut lexer = Lexer::new(input);
@@ -27,8 +27,8 @@ mod test {
         perform("13 + 13", Val::Int(13 + 13));
         perform(" 2 * 13 + 13", Val::Int(2 * 13 + 13));
         perform(" 2 * (13 + 13)", Val::Int(2 * (13 + 13)));
-        perform(" 2 * (13 + 13) / 2", Val::Int(2 * (13 + 13)/ 2));
-        perform("(2 + 2) * 10 ^ 2", Val::Int((2 + 2) * 10i128.pow(2)));
+        perform(" 2 * (13 + 13) / 2", Val::Int(2 * (13 + 13) / 2));
+        perform("(2 + 2) * 10 ^ 2", Val::Int((2 + 2) * 10i64.pow(2)));
         perform("(2 + 2) * 10.0", Val::Float((2 + 2) as f64 * 10.0));
     }
 }
